requirements.txt
gitpython>=3.1.40
pyyaml>=6.0.1
pytest>=7.4.3
fastapi>=0.105.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.2
